feat: Add Faust consumer to process Kafka messages
A new Python file, "faust_consumer.py", was added to the project in the "src/lesson-01/consumer" directory. This file includes a Faust consumer that listens to the Kafka broker located at "localhost:9092". It is set to receive messages from the topic 'my-topic' and prints the text of each incoming message.