feat: Specify schema version in create_message_class function
The create_message_class function has been updated to take the version of the schema we want to use. This modification was made in the faust_producer_registry.py file, in which 'schema_version' was added as a parameter, and now it's passed as an argument to the create_message_class function. This ensures we are using the correct version of the schema when creating our message class.